refactor api sync listener init
