+++ packages/api-sync/source/identifiers.ts
@@ -1,0 +1,3 @@
+export const Identifiers = {
+    Listeners: Symbol.for("API<Sync.Listeners>"),
+};
