+++ packages/api-sync/source/listeners/abstract-listener.ts
@@ -110,8 +110,7 @@ export abstract class AbstractListener<TEventData, TEntity extends { [key: strin
 			const entityRepository = this.makeEntityRepository(entityManager);
 
 			this.logger.debug(
-				`syncing ${entityRepository.metadata.tableNameWithoutPrefix} to database (added: ${
-					this.#addedEvents.size
+				`syncing ${entityRepository.metadata.tableNameWithoutPrefix} to database (added: ${this.#addedEvents.size
 				} removed: ${this.#removedEvents.size}))`,
 			);
