+++ packages/api-sync/source/service-provider.ts
@@ -1,50 +1,20 @@
-import { interfaces } from "@mainsail/container";
 import { Identifiers } from "@mainsail/contracts";
 import { Providers } from "@mainsail/kernel";
 
-import * as ApiSyncContracts from "./contracts";
-import { Mempool } from "./listeners/mempool";
-import { Peers } from "./listeners/peers";
-import { Plugins } from "./listeners/plugins";
+import { Identifiers as ApiSyncIdentifiers } from "./identifiers";
+import { Listeners } from "./listeners";
 import { Sync } from "./sync";
 
 export class ServiceProvider extends Providers.ServiceProvider {
-	#listeners: ApiSyncContracts.EventListener[] = [];
-
 	public async register(): Promise<void> {
+		this.app.bind(ApiSyncIdentifiers.Listeners).to(Listeners).inSingletonScope();
 		this.app.bind(Identifiers.ApiSync).to(Sync).inSingletonScope();
 
 		// Listen to events during register, so we can catch all boot events.
-		await this.#registerListeners();
-	}
-
-	public async boot(): Promise<void> {
-		await this.#bootListeners();
+		await this.app.get<Listeners>(ApiSyncIdentifiers.Listeners).register();
 	}
 
 	public async dispose(): Promise<void> {
-		await this.#disposeListeners();
-	}
-
-	async #registerListeners(): Promise<void> {
-		for (const constructor of [Peers, Plugins, Mempool]) {
-			const listener = this.app.resolve(constructor as interfaces.Newable<ApiSyncContracts.EventListener>);
-			await listener.register();
-			this.#listeners.push(listener);
-		}
-	}
-
-	async #bootListeners(): Promise<void> {
-		for (const listener of this.#listeners) {
-			await listener.boot();
-		}
-	}
-
-	async #disposeListeners(): Promise<void> {
-		for (const listener of this.#listeners) {
-			await listener.dispose();
-		}
-
-		this.#listeners = [];
+		await this.app.get<Listeners>(ApiSyncIdentifiers.Listeners).dispose();
 	}
 }
