+++ packages/api-sync/source/sync.ts
@@ -8,6 +8,8 @@ import { Contracts, Identifiers } from "@mainsail/contracts";
 import { Providers, Types, Utils } from "@mainsail/kernel";
 import { sleep } from "@mainsail/utils";
 import { performance } from "perf_hooks";
+import * as ApiSyncContracts from "./contracts";
+import { Identifiers as ApiSyncIdentifiers } from "./identifiers";
 
 interface DeferredSync {
 	block: Models.Block;
@@ -76,9 +78,11 @@ export class Sync implements Contracts.ApiSync.ISync {
 	private readonly createQueue!: Types.QueueFactory;
 	#queue!: Contracts.Kernel.Queue;
 
+	@inject(ApiSyncIdentifiers.Listeners)
+	private readonly listeners!: ApiSyncContracts.Listeners;
+
 	public async prepareBootstrap(): Promise<void> {
 		await this.migrations.run();
-
 		await this.#resetDatabaseIfNecessary();
 	}
 
@@ -86,6 +90,8 @@ export class Sync implements Contracts.ApiSync.ISync {
 		await this.#bootstrapConfiguration();
 		await this.#bootstrapState();
 		await this.#bootstrapTransactionTypes();
+
+		await this.listeners.bootstrap();
 
 		this.#queue = await this.createQueue();
 		await this.#queue.start();
@@ -154,8 +160,8 @@ export class Sync implements Contracts.ApiSync.ISync {
 
 			...(Utils.roundCalculator.isNewRound(header.height + 1, this.configuration)
 				? {
-						validatorRound: this.#createValidatorRound(header.height + 1),
-				  }
+					validatorRound: this.#createValidatorRound(header.height + 1),
+				}
 				: {}),
 		};
